# swagger/docApiDuvidas.yaml
openapi: 3.0.1
info:
  title: Forum do programa de bolsas.
  description: >-
    API para postagem de dúvidas e respectivas soluções, dentro do programa de
    bolsas da empresa Compasso/UOL.
  version: 1.0.0
  termsOfService: https://compassouol.com
  contact:
    name: Suporte a desenvolvedores
    email: compasso@hotmail.com
  license:
    name: 'Licença: GPLv3'
    url: https://www.gnu.org/licenses/gpl-3.0.html
servers:
  - url: http://localhost:8080/
    description: API de teste
paths:
...
  /turmas:
    get:
      tags:
        - turmas
      summary: Recupera uma lista de turmas.
      description: Pode recuperar todas as turmas cadastradas
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TurmasDTO'
    post:
      tags:
        - turmas
      summary: Cadastra uma nova turma
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TurmaFormDTO'
      responses:
        '201':
          description: Cadastrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TurmaDTO'
  /turmas/{id}:
    get:
      tags:
        - turmas
      summary: Recupera uma turma pelo Id
      parameters:
        - name: id
          in: path
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TurmaDTO'
        '404':
          description: Turma não encontrada
          content:
            application/json:
              example: Not Found
    put:
      tags:
        - turmas
      summary: Atualiza a turma de determinado ID
      parameters:
        - name: id
          in: path
          schema:
            type: integer
          required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TurmaFormDTO'
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TurmaDTO'
        '404':
          description: Turma não encontrada
    delete:
      tags:
        - turmas
      summary: Remove a turma de determinado ID
      parameters:
        - name: id
          in: path
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Sucesso
        '404':
          description: Turma não encontrada
  /turmas/{id}/sprints:
    get:
      tags:
        - turmas
      summary: Recupera todas sprints de uma determinada turma
      parameters:
        - name: id
          in: path
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TurmasDTO'
        '404':
          description: Turma não encontrada
    post:
      tags:
        - turmas
      summary: Adiciona uma sprint a uma turma
      parameters:
        - name: id
          in: path
          schema:
            type: integer
          required: true
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TurmaAddSprintFormDTO'
      responses:
        '201':
          description: Cadastrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SprintDTO'
        '404':
          description: Turma não encontrada
  /turmas/{id1}/sprints/{id2}:
    delete:
      tags:
        - turmas
      summary: Remove uma determinada sprint de uma determinada turma
      parameters:
        - name: id1
          in: path
          schema:
            type: integer
          required: true
        - name: id2
          in: path
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Sucesso
        '404':
          description: Turma ou sprint não encontrada
  /turmas/{id1}/usuarios/{id2}:
    post:
      tags:
        - turmas
      summary: Adiciona um usuario a uma determinada turma
      parameters:
        - name: id1
          in: path
          schema:
            type: integer
          required: true
        - name: id2
          in: path
          schema:
            type: integer
          required: true
      responses:
        '201':
          description: Cadastrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TurmaDTO'
        '404':
          description: Turma ou usuario não encontrados
    delete:
      tags:
        - turmas
      summary: Remove um determinado usuario de uma determinada turma
      parameters:
        - name: id1
          in: path
          schema:
            type: integer
          required: true
        - name: id2
          in: path
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Sucesso
        '404':
          description: Turma ou usuario não encontrada
  /turmas/{id}/usuarios:
    get:
      tags:
        - turmas
      summary: Recupera todos os usuarios de uma determinada turma
      parameters:
        - name: id
          in: path
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsuariosDTO'
        '404':
          description: Turma não encontrada
...
security:
  - auth: []
components:
  securitySchemes:
    auth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TurmaDTO:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nome:
          type: string
          example: Spring Boot
        usuariosIds:
          type: array
          items:
            type: integer
            example: 1
        sprints:
          type: array
          items:
            type: string
            example: 'Id: 3 Título: Primeira Sprint'
    TurmasDTO:
      type: array
      items:
        $ref: '#/components/schemas/TurmaDTO'
...
    SprintDTO:
      type: object
      properties:
        id:
          type: integer
          example: 3
        titulo:
          type: string
          example: Primeira Sprint
...
    TurmaAddSprintFormDTO:
      type: array
      items:
        type: object
        properties:
          sprintsIds:
            type: integer
            example: 4
...
    TurmaFormDTO:
      type: object
      properties:
        nome:
          type: string
          example: Nome da turma
    UsuarioDTO:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nome:
          type: string
          example: João da Silva
        email:
          type: string
          example: rosrigo@gmail.com
        perfilUsuario:
          type: string
          example: usuario
    UsuariosDTO:
      type: array
      items:
        $ref: '#/components/schemas/UsuarioDTO'
...